MGDCTRS-1667: feat: update
