+++ src/apis/api.ts
@@ -62,10 +62,6 @@ type ConnectorNamespaceProps = {
   namespaceId: string;
 } & CommonApiProps;
 
-type ConnectorClusterProps = {
-  clusterId: string;
-} & CommonApiProps;
-
 type ConnectorTypeProps = {
   connectorTypeId: string;
 } & CommonApiProps;
@@ -382,11 +378,10 @@ export const registerEvalNamespace = ({
   };
 };
 
-export const getCluster = ({
+export const getClusters = ({
   accessToken,
   connectorsApiBasePath,
-  clusterId,
-}: ConnectorClusterProps) => {
+}: CommonApiProps) => {
   const clusterAPI = new ConnectorClustersApi(
     new Configuration({
       accessToken,
@@ -398,7 +393,7 @@ export const getCluster = ({
     const CancelToken = axios.CancelToken;
     const source = CancelToken.source();
     clusterAPI
-      .getConnectorCluster(clusterId, {
+      .listConnectorClusters(undefined, undefined, undefined, undefined, {
         cancelToken: source.token,
       })
       .then((response) => {
@@ -407,7 +402,7 @@ export const getCluster = ({
       .catch((error) => {
         if (!axios.isCancel(error)) {
           console.log(
-            `error response fetching connector cluster: ${clusterId}`,
+            `error response fetching connector clusters}`,
             error.response.data.reason
           );
         }
@@ -486,7 +481,9 @@ export const fetchConnectorNamespaces = ({
       .map((s) => `(${s})`)
       .join(' AND ');
     namespacesAPI
-      .listConnectorNamespaces(`${page}`, `${size}`, undefined, searchString)
+      .listConnectorNamespaces(`${page}`, `${size}`, undefined, searchString, {
+        cancelToken: source.token,
+      })
       .then((response) => {
         onSuccess({
           items: response.data.items || [],
